Move Furniture Signpost Resource methods to a Generic Base Resource
This allows using it as a base when creating more Resources in the future.
